macros: promote impl_wire_enum crate-wide + dedupe encoder/test infra (typed-status sweep PR 4a)
Zero behavior change. Sets up PR 4b (ExecutionFilterSide) and PR 5b
(ExecutionSide) to use the same shape without further infra work.

- Move impl_str_partial_eq! and impl_wire_enum! from contracts/types.rs
  to a new src/macros.rs reachable crate-wide via #[macro_use] mod
  macros; in lib.rs. Project convention had been module-local macros;
  justified by 6-consumer count (OptionRight, SecurityIdType, LegAction
  in contracts/; OrderStatusKind retrofit here + imminent
  ExecutionFilterSide + ExecutionSide in orders/). Orphan rule
  forecloses blanket-impl alternatives (rule 25 case a).

- Promote check_wire_enum_round_trip and check_wire_enum_rejects_unknown
  from contracts/types_tests.rs to a new pub mod wire_enum in
  common/test_utils.rs. Generic helper functions, not a macro (rule 25
  preferred shape). orders/tests.rs and contracts/types_tests.rs both
  consume.

- Add some_display(Option<and impl Display>) returning Option<String>
  next to some_str in proto/encoders.rs and convert the two existing
  PR 2 / PR 3b sites (right, sec_id_type). Third site (PR 4b
  filter.side) hits the rule-of-three tripwire; closes the PR 3b
  /simplify tracked deferral.

- Retrofit OrderStatusKind in orders/mod.rs to as_str + from_wire +
  impl_wire_enum!. Drops the hand-rolled Display + FromStr; gains
  ToField (additive, no removal). Closes the PR 3a follow-up note.
